# Upgrade SIMDUTF, 8x faster atob