--- a/src/bun.js/bindings/Base64Helpers.h
+++ b/src/bun.js/bindings/Base64Helpers.h
@@ -1,0 +1,10 @@
+#include "root.h"
+
+namespace Bun {
+
+namespace Base64 {
+
+WebCore::ExceptionOr<WTF::String> atob(const WTF::String& encodedString);
+
+}
+}