… src/bun.js/bindings/simdutf.cpp
… src/bun.js/bindings/simdutf.h