### What does this PR do?

Fixes #202 

Thanks to @lemire


New:
```ts
cpu: Apple M3 Max
runtime: bun 1.1.9 (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
atob(32 chars)        32.94 ns/iter  (29.98 ns … 106.67 ns)  33.05 ns  41.25 ns  47.05 ns
atob(512 chars)      133.53 ns/iter (106.74 ns … 204.08 ns) 148.22 ns 194.12 ns 202.67 ns
atob(65536 chars)      8.04 µs/iter   (6.46 µs … 689.79 µs)   6.92 µs  12.04 µs  14.38 µs
atob(524288 chars)    63.59 µs/iter  (51.54 µs … 690.67 µs)  61.83 µs  368.5 µs 402.92 µs
atob(8388608 chars)    1.09 ms/iter   (825.71 µs … 1.65 ms)   1.24 ms   1.51 ms   1.53 ms
```

Bun v1.1.7:
```ts
cpu: Apple M3 Max
runtime: bun 1.1.7 (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
atob(32 chars)        66.46 ns/iter  (58.37 ns … 135.62 ns)  67.25 ns  93.29 ns 101.02 ns
atob(512 chars)      720.59 ns/iter (659.85 ns … 808.02 ns)  741.9 ns 808.02 ns 808.02 ns
atob(65536 chars)     82.61 µs/iter  (76.13 µs … 536.42 µs)  82.75 µs    101 µs  105.5 µs
atob(524288 chars)   655.23 µs/iter   (608.42 µs … 1.28 ms) 662.67 µs 934.96 µs 973.83 µs
atob(8388608 chars)   10.39 ms/iter    (9.82 ms … 12.09 ms)   10.6 ms  12.09 ms  12.09 ms
```

Deno:
```ts
cpu: unknown
runtime: deno 1.43.3 (aarch64-apple-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
atob(32 chars)        60.82 ns/iter  (55.07 ns … 149.17 ns)  60.55 ns  68.07 ns  78.21 ns
atob(512 chars)      180.15 ns/iter  (167.72 ns … 208.7 ns) 183.82 ns 201.11 ns 205.44 ns
atob(65536 chars)     16.74 µs/iter   (15.13 µs … 96.33 µs)  16.33 µs  52.08 µs  55.25 µs
atob(524288 chars)   176.97 µs/iter (136.75 µs … 363.38 µs) 178.92 µs 213.58 µs 237.38 µs
atob(8388608 chars)     2.3 ms/iter     (2.25 ms … 2.45 ms)   2.33 ms   2.45 ms   2.45 ms
```

Node $HEAD:
```ts
cpu: Apple M3 Max
runtime: node v23.0.0-head (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
atob(32 chars)        45.04 ns/iter   (41.45 ns … 59.56 ns)  44.75 ns  53.47 ns  55.94 ns
atob(512 chars)      116.53 ns/iter (105.88 ns … 138.93 ns) 117.51 ns 127.24 ns 129.55 ns
atob(65536 chars)      9.95 µs/iter    (8.5 µs … 144.25 µs)   9.54 µs  48.13 µs  50.13 µs
atob(524288 chars)   136.06 µs/iter  (94.38 µs … 270.13 µs) 139.08 µs 158.42 µs 168.54 µs
atob(8388608 chars)    1.92 ms/iter     (1.79 ms … 2.15 ms)   1.96 ms    2.1 ms   2.13 ms

```

Node v18:
```ts
cpu: Apple M3 Max
runtime: node v18.20.2 (arm64-darwin)

benchmark                time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------- -----------------------------
atob(32 chars)        682.5 ns/iter (656.21 ns … 705.69 ns) 687.66 ns 705.69 ns 705.69 ns
atob(512 chars)        8.12 µs/iter     (8.05 µs … 8.34 µs)   8.15 µs   8.34 µs   8.34 µs
atob(65536 chars)      1.05 ms/iter   (958.38 µs … 1.26 ms)   1.06 ms   1.19 ms   1.24 ms
atob(524288 chars)     8.42 ms/iter     (7.99 ms … 8.97 ms)   8.45 ms   8.97 ms   8.97 ms
atob(8388608 chars)  133.87 ms/iter (130.46 ms … 141.06 ms) 135.31 ms 141.06 ms 141.06 ms
```

### How did you verify your code works?

Ran existing tests